move runUfsTests
